+++ gatling-core/src/main/scala/com/excilys/ebi/gatling/core/result/reader/GeneralStats.scala
@@ -18,7 +18,7 @@ package com.excilys.ebi.gatling.core.result.reader
 import com.excilys.ebi.gatling.core.result.reader.DataReader.NO_PLOT_MAGIC_VALUE
 
 object GeneralStats {
-	val NO_PLOT = GeneralStats(NO_PLOT_MAGIC_VALUE, NO_PLOT_MAGIC_VALUE, 0L, NO_PLOT_MAGIC_VALUE, NO_PLOT_MAGIC_VALUE, NO_PLOT_MAGIC_VALUE, NO_PLOT_MAGIC_VALUE, NO_PLOT_MAGIC_VALUE)
+	val NO_PLOT = GeneralStats(NO_PLOT_MAGIC_VALUE, NO_PLOT_MAGIC_VALUE, 0, NO_PLOT_MAGIC_VALUE, NO_PLOT_MAGIC_VALUE, NO_PLOT_MAGIC_VALUE, NO_PLOT_MAGIC_VALUE, NO_PLOT_MAGIC_VALUE)
 }
 
-case class GeneralStats(min: Long, max: Long, count: Long, mean: Long, stdDev: Long, percentile1: Long, percentile2: Long, meanRequestsPerSec: Long)
+case class GeneralStats(min: Int, max: Int, count: Int, mean: Int, stdDev: Int, percentile1: Int, percentile2: Int, meanRequestsPerSec: Int)
